finished modal for more info
